requirements.txt
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==0.5.2
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
autograd==1.6.2
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
bokeh==3.1.1
boltons==23.1.1
boto3==1.34.19
botocore==1.34.19
bravado==11.0.3
bravado-core==6.1.1
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1648883617327/work
cachetools==5.3.3
certifi @ file:///croot/certifi_1700501669400/work/certifi
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
colorama==0.4.6
comm==0.2.0
contourpy==1.1.1
crcmod==1.7
cryptography==41.0.4
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
dataclasses==0.6
datasets==2.19.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
dnspython==2.6.1
docker-pycreds==0.4.0
e2cnn==0.2.3
emmet-core==0.68.0
escnn @ file:///home/yang1683/Research/SymDet/SymmetryDetection/escnn
exceptiongroup==1.1.3
executing==2.0.1
fairscale==0.4.13
fastjsonschema==2.18.1
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1698714947081/work
filetype==1.2.0
fire==0.5.0
flatten-dict==0.4.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.2.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.37
huggingface-hub==0.22.2
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
imageio==2.31.5
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.12.3
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1655912915439/work
jmespath==0.10.0
joblib==1.3.2
json2html==1.3.0
json5==0.9.14
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.1
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
latexcodec==3.0.0
lazy_loader==0.3
lie-learn==0.0.1.post1
line-profiler==4.1.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1610127565888/work
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-base_1594091694890/work
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mmcv-full==1.7.0
model-index==0.1.11
monotonic==1.6
monty==2023.9.25
mp-api==0.35.1
mpcontribs-client==5.7.0
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1678228039184/work
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1667391556261/work
notebook==7.0.6
notebook_shim==0.2.3
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opencv-python-headless==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.26
ordered-set==4.1.0
oss2==2.17.0
overrides==7.4.0
packaging==23.2
palettable==3.3.3
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///croot/pillow_1695134008276/work
Pint==0.19.2
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
plotly==5.19.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py3nj==0.2.1
pyarrow==16.0.0
pyarrow-hotfix==0.6
pybtex==0.24.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.14
Pygments==2.16.1
pyIsEmail==2.0.1
pymanopt==2.2.0
pymatgen==2023.8.10
pymongo==4.6.2
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1690737849915/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-json-logger==2.0.7
pytorch-transformers==1.2.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.0
QtPy==2.4.1
referencing==0.30.2
regex==2023.10.3
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
requests-futures==1.0.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.12.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
sacremoses==0.1.1
safetensors==0.4.0
scikit-image==0.21.0
scipy==1.9.3
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.31.0
setproctitle==1.3.3
shapely==2.0.1
simplejson==3.19.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spglib==2.3.1
stack-data==0.6.3
swagger-spec-validator==3.0.3
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1684180539862/work
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
terminado==0.17.1
tifffile==2023.7.10
tinycss2==1.2.1
tokenizers==0.14.1
tomli==2.0.1
torch==1.12.0+cu113
torchaudio==0.12.0+cu113
torchvision==0.13.0+cu113
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827257044/work
tqdm==4.65.2
traitlets==5.13.0
transformers==4.35.0
triton==2.1.0
types-python-dateutil==2.8.19.14
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
ujson==5.9.0
uncertainties==3.1.7
uri-template==1.3.0
urllib3==1.26.17
uuid==1.30
wandb==0.15.11
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
xgboost==2.0.3
xxhash==3.4.1
xyzservices==2023.10.1
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
